Moved container components into their own directory and added a PageNotFound component
