+++ src/containers/StockDetails/StockDetails.module.scss
