+++ src/containers/StocksIndex/StocksIndex.module.scss
